+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 v1.6.3
+ - 感谢@zhou-a-xing 调整匿名结构体字段顺序
  - 感谢@无泪 反馈的问题,避免IEntityMap默认实现IEntityStruct接口
  - 感谢@cucuy 对www.zorm.cn官网的修改
  - 完善文档,注释